requirements.txt
# ── Core Framework ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# ── Database ──
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1

# ── Job Queue ──
celery[redis]==5.4.0
redis==5.2.1

# ── AI / LLM ──
openai==1.58.1
httpx==0.28.1

# ── TTS ──
TTS==0.22.0
elevenlabs==1.15.0
edge-tts==7.0.2

# ── Image Generation ──
Pillow==11.1.0

# ── Subtitle ──
openai-whisper==20240930
mutagen==1.47.0

# ── YouTube Upload ──
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# ── Utilities ──
aiofiles==24.1.0
python-multipart==0.0.20

# ── Testing ──
pytest==8.3.4
pytest-asyncio==0.25.0

# ── Monitoring ──
flower==2.0.1
